refact: restructure tray creation
